fix(suite): onboarding last step with BTC-only
